refactor: split tab_executor_test.go into 3 files and clean up .gitignore
- Split tab_executor_test.go into:
  - tab_executor_test.go (unit tests + edge cases)
  - tab_executor_integration_test.go (integration, stress, TabManager tests)
  - tab_executor_benchmark_test.go (all benchmarks)
- Remove individual test output files from .gitignore (use dist/ instead)
- Move test output files to dist/ folder
